+++ mycodo/mycodo_flask/authentication_routes.py
@@ -238,7 +238,7 @@ def newremote():
         if User().check_password(
                 pass_word, user.password_hash) == user.password_hash:
             try:
-                with open('/var/www/mycodo/mycodo/mycodo_flask/ssl_certs/cert.pem', 'r') as cert:
+                with open('/var/mycodo-root/mycodo/mycodo_flask/ssl_certs/cert.pem', 'r') as cert:
                     certificate_data = cert.read()
             except Exception:
                 certificate_data = None