+++ CHANGELOG.md
@@ -10,6 +10,12 @@
  - Fix starting web server during install
  - Fix issue with gunicorn worker timeouts on Raspberry Pi Zeros ([#365](https://github.com/kizniche/mycodo/issues/365))
  - Fix command variable replacement for Output conditionals ([#367](https://github.com/kizniche/mycodo/issues/367))
+ - Fix pH Input causing an error with a deactivated Calibration Measurement ([#369](https://github.com/kizniche/mycodo/issues/369))
+
+### Miscellaneous
+
+ - Move mycodo root symlink from /var/www to /var
+ - Create symlinks in PATH for mycodo-backup, mycodo-client, mycodo-commands, mycodo-daemon, mycodo-pip, mycodo-python, mycodo-restore, and mycodo-wrapper
 
 
 ## 5.5.3 (2017-12-29)